+++ src/Sable.Tools/RasterStateCommand.cs
@@ -10,22 +10,22 @@ namespace Sable.Tools;
 /// </summary>
 public readonly struct RasterState
 {
-    public readonly byte[] Pixels;
-    public readonly byte[]? Mask;   // layer-aligned mask snapshot (ExpandToCover/TrimToContent reallocs it too)
+    public readonly float[] Pixels;   // RGBA32F (the bit-depth float pipeline, PLAN §6)
+    public readonly byte[]? Mask;   // layer-aligned 8-bit mask snapshot (ExpandToCover/TrimToContent reallocs it too)
     public readonly int Width, Height, OffsetX, OffsetY;
 
-    private RasterState(byte[] pixels, byte[]? mask, int w, int h, int ox, int oy)
+    private RasterState(float[] pixels, byte[]? mask, int w, int h, int ox, int oy)
     { Pixels = pixels; Mask = mask; Width = w; Height = h; OffsetX = ox; OffsetY = oy; }
 
     /// <summary>Capture a (deep-copied) snapshot of the layer's current raster state.</summary>
     public static RasterState Capture(PixelLayer layer)
-        => new((byte[])layer.Pixels.Clone(), layer.Mask is { } m ? (byte[])m.Clone() : null,
+        => new((float[])layer.Pixels.Clone(), layer.Mask is { } m ? (byte[])m.Clone() : null,
                layer.Width, layer.Height, layer.OffsetX, layer.OffsetY);
 
     /// <summary>Restore this state into the layer (fresh buffer copy so the snapshot stays immutable).</summary>
     public void ApplyTo(PixelLayer layer)
     {
-        layer.SetBuffer(Width, Height, (byte[])Pixels.Clone());
+        layer.SetBuffer(Width, Height, (float[])Pixels.Clone());
         layer.OffsetX = OffsetX;
         layer.OffsetY = OffsetY;
         layer.Mask = Mask is { } m ? (byte[])m.Clone() : null;
